Time: 5 ms (58.49%) | Memory: 16.6 MB (13.65%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        int prod=1;
-        int res=INT_MIN;
+         double prod=1;
+         double res=INT_MIN;
         for(int i=0;i<nums.size();i++){
             prod=prod*nums[i];
             //prod=max(prod,nums[i]);
